pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apirunner"
version = "0.1.0"
description = "Minimal HTTP API test runner (MVP)"
requires-python = ">=3.10"
authors = [{ name = "APIRunner Team" }]
readme = "README.md"
dependencies = [
  "httpx>=0.27",
  "pydantic>=2.6",
  "jinja2>=3.1",
  "jmespath>=1.0",
  "PyYAML>=6.0",
  "rich>=13.7",
  "typer>=0.12",
]

[project.scripts]
arun = "apirunner.cli:app"

